SVNFileUtil.rename - in no-jna fallback mode on Windows do not delete target of rename when rename is case-only.

+++ svnkit/src/main/java/org/tmatesoft/svn/core/internal/wc/SVNFileUtil.java
@@ -703,9 +703,12 @@ public class SVNFileUtil {
                 renamed = true;
             } 
             if (!renamed) {
+            	boolean caseOnly = dst.getAbsolutePath().equalsIgnoreCase(src.getAbsolutePath());
                 boolean wasRO = dst.exists() && !dst.canWrite();
                 setReadonly(src, false);
-                setReadonly(dst, false);
+                if (!caseOnly) {
+                	setReadonly(dst, false);
+                }
                 // try simple atomic rename first
                 if (src.renameTo(dst)) {
                     return;
@@ -713,7 +716,9 @@ public class SVNFileUtil {
                 // use special loop on windows.
                 long sleep = 1;
                 for (int i = 0; i < FILE_CREATION_ATTEMPTS_COUNT; i++) {
-                    dst.delete();
+                	if (!caseOnly) {
+                		dst.delete();
+                	}
                     if (src.renameTo(dst)) {
                         if (wasRO && !isOpenVMS) {
                             dst.setReadOnly();
